+++ aligenie-oauth2_1_0/ChangeLog.md
@@ -1,3 +1,11 @@
+2025-09-18 Version: 2.0.0
+- Support API ExecuteSmartHomeScene.
+- Support API GetSmartHomeSceneList.
+- Update API CreatePlayingList: add request parameters OpenCreatePlayingListRequest.RequestId.
+- Update API CreatePlayingList: delete response parameters Body.Code.
+- Update API CreatePlayingList: delete response parameters Body.Message.
+
+
 2025-05-14 Version: 1.1.1
 - Update API CreatePlayingList: add response parameters Body.Code.
 - Update API CreatePlayingList: add response parameters Body.Message.